Added profile inputs
Added profile inputs on registration page and attempted to add it to profile when already logged in
Heads up! registering does not throw an error and since I added all of this you might have to register again, but attempting to look at profile after logging in or registering throws an error.